refactor: separate KeypadWrapper
