+++ src/domains/apps/calculator/views/Calculator/Calculator.css.ts
@@ -1,5 +1,5 @@
 import { style } from '@vanilla-extract/css';
-import { calcColorTokens } from './calcTheme.css';
+import { calcColorTokens } from '../theme.css';
 
 export const movableArea = style({
   display: 'flex',
@@ -15,12 +15,4 @@ export const displayContainer = style({});
 export const displayText = style({
   fontSize: 24,
   textAlign: 'right',
-});
-
-export const keypadContainer = style({
-  display: 'grid',
-  gridTemplateColumns: '56px 57px 57px 59px',
-  gridTemplateRows: 'repeat(5, 47px)',
-  gap: '1px',
-  backgroundColor: calcColorTokens.grey700,
 });
