+++ src/domains/apps/calculator/views/KeypadWrapper/index.ts
@@ -1,0 +1,1 @@
+export * from './KeypadWrapper';
